Ran prettier to format files and renamed populate files
